+++ src/USER-DPD/fix_eos_table.cpp
@@ -116,11 +116,11 @@ void FixEOStable::init()
   } else {
     for (int i = 0; i < nlocal; i++)
       if (mask[i] & groupbit) {
-        if(dpdTheta[i] <= double(0.0))
+        if(dpdTheta[i] <= 0.0)
           error->one(FLERR,"Internal temperature <= zero");
         energy_lookup(dpdTheta[i],tmp);
-        uCond[i] = tmp / double(2.0);
-        uMech[i] = tmp / double(2.0);
+        uCond[i] = tmp / 2.0;
+        uMech[i] = tmp / 2.0;
       }
   }
 }
@@ -138,7 +138,7 @@ void FixEOStable::post_integrate()
   for (int i = 0; i < nlocal; i++)
     if (mask[i] & groupbit){
       temperature_lookup(uCond[i]+uMech[i],dpdTheta[i]);
-      if(dpdTheta[i] <= double(0.0))
+      if(dpdTheta[i] <= 0.0)
         error->one(FLERR,"Internal temperature <= zero");
     }
 }
@@ -156,7 +156,7 @@ void FixEOStable::end_of_step()
   for (int i = 0; i < nlocal; i++)
     if (mask[i] & groupbit){
       temperature_lookup(uCond[i]+uMech[i],dpdTheta[i]);
-      if(dpdTheta[i] <= double(0.0))
+      if(dpdTheta[i] <= 0.0)
         error->one(FLERR,"Internal temperature <= zero");
     }
 }
